+++ packages/iov-tendermint-rpc/src/requests.ts
@@ -122,7 +122,7 @@ export interface SubscribeRequest {
   readonly method: Method.Subscribe;
   readonly query: {
     readonly type: SubscriptionEventType;
-    readonly tags?: ReadonlyArray<QueryTag>;
+    readonly raw?: QueryString;
   };
 }